refactor: 코드베이스 컨벤션 준수 및 공통 로직 모듈화 (#50) - DesignDetailCard/ThemeDetailCard → 공통 BoardDetailCard로 통합 - 스와이프 로직 useVerticalSwipe, 댓글 생성 useCreateComment 훅 분리 - MyPage 하위 컴포넌트(ActivityTab, ThemeGridTab, MyPageBoardCard) 파일 분리 - Login/SignUp Page → Form 컴포넌트 분리, Preview 컴포넌트 분리 - ITag 중복 제거, IMoreMenuItem types.ts 이동, IBoardDetailBase 공통 타입 추가 - apiClient.ts → ApiClient.ts 리네임, const tabs → TABS 상수명 수정 - scrollbar-hidden 커스텀 CSS 제거 → Tailwind arbitrary values 교체 - CLAUDE.md 상태 변수 네이밍 camelCase로 수정
