+++ src/index.css
@@ -25,16 +25,3 @@
   }
 }
 
-/**
-* 공통 유틸리티 클래스
-*/
-@layer utilities {
-  .scrollbar-hidden {
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-  }
-
-  .scrollbar-hidden::-webkit-scrollbar {
-    display: none;
-  }
-}
