+++ src/pages/design-community/List.tsx
@@ -5,7 +5,7 @@ import TabMenu from "../../components/common/TabMenu";
 import type { TabId } from "../../types/community/theme";
 import { useDesignBoards } from "../../services/hooks/design/useDesignBoards";
 
-const tabs: { id: TabId; label: string }[] = [
+const TABS: { id: TabId; label: string }[] = [
   { id: 'activity', label: '활동' },
   { id: 'keyword', label: '키워드' },
 ];
@@ -16,7 +16,7 @@ export default function List() {
 
   return (
     <div>
-      <TabMenu tabs={tabs} activeTab={activeTab} onTabChange={setActiveTab} />
+      <TabMenu tabs={TABS} activeTab={activeTab} onTabChange={setActiveTab} />
       <div className="flex-1">
         {activeTab === 'activity' && (
           <DesignActivityTab
