+++ src/pages/theme-community/BoardWrite.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useBoardWrite } from '../../services/hooks/theme/useBoardWrite';
 import BoardWriteForm from '../../components/community/BoardWriteForm';
-import Text from '../../components/common/Text';
+import ThemeBoardPreview from '../../components/community/theme/ThemeBoardPreview';
 import type { IUserTheme } from '../../types/community/theme';
 
 interface IBoardWriteLocationState {
@@ -21,48 +21,22 @@ export default function BoardWrite() {
   }, [state, navigate]);
 
   const selectedTheme = state?.selectedTheme;
-
   const [currentImageIndex, setCurrentImageIndex] = useState(0);
-  const imageCount = selectedTheme?.images.length ?? 1;
 
   const { handleSubmit, ...formProps } = useBoardWrite(selectedTheme!);
 
   if (!selectedTheme) return null;
-
-  const MAX_DOTS = 5;
-  const total = Math.max(imageCount, 1);
-  const visibleCount = Math.min(MAX_DOTS, total);
-  const startIndex = Math.min(
-    Math.max(currentImageIndex - Math.floor(MAX_DOTS / 2), 0),
-    Math.max(total - MAX_DOTS, 0),
-  );
 
   return (
     <BoardWriteForm
       {...formProps}
       onSubmit={handleSubmit}
       preview={
-        <div className="mb-5 flex flex-col items-center">
-          <div className="flex h-44 w-full items-center justify-center rounded-xl bg-secondary-200">
-            <Text variant="REGULAR_14" className="text-secondary-400">
-              {selectedTheme.themeName}
-            </Text>
-          </div>
-          <div className="mt-2 flex gap-1.5">
-            {Array.from({ length: visibleCount }, (_, i) => {
-              const dotIndex = startIndex + i;
-              const isActive = dotIndex === currentImageIndex;
-              return (
-                <button
-                  key={dotIndex}
-                  type="button"
-                  onClick={() => setCurrentImageIndex(dotIndex)}
-                  className={`h-1.5 w-1.5 rounded-full transition-colors ${isActive ? 'bg-primary' : 'bg-secondary-300'}`}
-                />
-              );
-            })}
-          </div>
-        </div>
+        <ThemeBoardPreview
+          selectedTheme={selectedTheme}
+          currentImageIndex={currentImageIndex}
+          onIndexChange={setCurrentImageIndex}
+        />
       }
     />
   );
