+++ src/pages/theme-community/List.tsx
@@ -4,7 +4,7 @@ import TabMenu from "../../components/common/TabMenu";
 import type { TabId } from "../../types/community/theme";
 import { useThemeBoards } from "../../services/hooks/theme/useThemeBoards";
 
-const tabs: { id: TabId; label: string }[] = [
+const TABS: { id: TabId; label: string }[] = [
   { id: 'activity', label: '활동' },
   { id: 'keyword', label: '키워드' },
 ];
@@ -15,7 +15,7 @@ export default function List() {
 
   return (
     <div>
-      <TabMenu tabs={tabs} activeTab={activeTab} onTabChange={setActiveTab} />
+      <TabMenu tabs={TABS} activeTab={activeTab} onTabChange={setActiveTab} />
       <div className="flex-1">
         {activeTab === 'activity' && (
           <ActivityTab
