+++ src/services/api/AuthService.ts
@@ -1,4 +1,4 @@
-import apiClient from './apiClient';
+import apiClient from './ApiClient';
 import type { IAuthTokens } from '../../types/auth/types';
 
 export const AuthService = {
